webp-load: use signed capable return type when returning -1

+++ operations/external/webp-load.c
@@ -105,7 +105,7 @@ read_from_stream (GInputStream *stream,
   return read;
 }
 
-static gsize
+static glong
 decode_from_stream (GInputStream *stream,
                     WebPIDecoder *decoder)
 {
